Remove factory usage to only use prometheus
